fix(git submodules): point to working submodules
I synked with upstream, but during the Gitkraken's adviced rebase I guess that the git submodules were overridden wrongly (opposite commit order due to the rebase). Going back to the latest working commit to see the SHA for the submodules and applying them on the latest commit again as a new commit solved it.